# prevent read to Redis from OrganizationMonitor MW